+++ package.json
@@ -29,9 +29,11 @@
     "url": "https://github.com/robertoachar/generator-node/issues"
   },
   "dependencies": {
+    "esm": "3.2.22",
     "yeoman-generator": "3.2.0"
   },
   "devDependencies": {
+    "@babel/preset-env": "7.4.3",
     "@types/jest": "24.0.11",
     "@types/yeoman-generator": "3.1.1",
     "coveralls": "3.0.3",
